헤더 파일 별도로 분리하기
